Alteração nos nomes dos controllers.
