--- itaretif/src/main/java/edu/ifrn/itaretif/controllers/LoginController.java
@@ -1,17 +1,0 @@
-package edu.ifrn.itaretif.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-@Controller
-	public class LoginController {
-	
-	@GetMapping("/login")
-	public String login() {
-		return "login";
-		
-	}
-	
-}
